pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "FastApi Boilerplate"
authors = ["Aayush Kurup <aayushkurup786@outlook.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fastapi_boilerplate"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
dependency-injector = "^4.41.0"
uvicorn = "^0.20.0"
orjson = "^3.8.5"
pydantic = {extras = ["email"], version = "^1.10.4"}
python-dotenv = "^0.21.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.10"
taskipy = "^1.10.3"

[tool.taskipy.tasks]
dev = "python main.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
